Add in a very, very hacked up version of pmcstat.c from usr.sbin/pmcstat in freebsd.
This is part refactor, part hack to get this thing into shape to spit out _something_.

It's temporary - I'll eventually replace it with something very minimal just for
watching a set of PMCs (counter mode, not sample mode) just for this particular
setup.  But it's enough to at least get the thing to work.